Use `usePlayerHandler` within `Battle` and childs
